+++ sections/job/model/job_type.php
@@ -11,6 +11,8 @@
  */
 
 include_once __DIR__ . "/../../../parametres/varmodtypegen/model/modelTypeGeneric.php";
+include_once __DIR__ . "/../../../parametres/model/model/model.php";
+include_once __DIR__ . "/../../../parametres/type/model/type.php";
 include_once __DIR__ . "/jobTypeParameter.php";
 
 class JobType extends \ModelTypeGeneric implements \JsonSerializable
@@ -27,8 +29,8 @@ class JobType extends \ModelTypeGeneric implements \JsonSerializable
      *
      * @param int $id The id of the JobType in the database
      * @param int $jobId The id of the Job to which this JobType belongs
-     * @param int $modelId The id of the Model associated with this JobType (the one that was modified)
-     * @param int $typeNo The importNumber of the Type associated with this JobType (the one that was modified)
+     * @param \Model $model The Model associated with this JobType (the one that was modified)
+     * @param \Type $type The Type associated with this JobType (the one that was modified)
      * @param string $mprFile The contents of the .mpr file associated to this JobType if not using a generic file
      * @param int $genericId The id of the Generic associated to this JobType
      * @param string $timestamp A timestamp of the last modification applied to this JobType (leave null)
@@ -39,11 +41,11 @@ class JobType extends \ModelTypeGeneric implements \JsonSerializable
      * @author Marc-Olivier Bazin-Maurice
      * @return \JobType
      */
-    public function __construct(?int $id = null, ?int $jobId = null, ?int $modelId = null, ?int $typeNo = null,
-        ?string $mprFile = null, ?int $genericId = null, ?string $timestamp = null, array $parameters = array(), 
+    public function __construct(?int $id = null, ?int $jobId = null, ?\Model $model = null, ?\Type $type = null,
+        ?string $mprFile = null, ?string $timestamp = null, array $parameters = array(), 
         array $parts = array())
     {
-        parent::__construct($modelId, $typeNo, $parameters);
+        parent::__construct($model, $type, $parameters);
         $this->setId($id);
         $this->setJobId($jobId);
         $this->setMprFile($mprFile);
@@ -67,7 +69,7 @@ class JobType extends \ModelTypeGeneric implements \JsonSerializable
         // Récupérer le test
         $stmt = $db->getConnection()->prepare(
             "SELECT `jt`.`job_id` AS `jobId`, `jt`.`door_model_id` AS `modelId`, `jt`.`type_no` AS `typeNo`, 
-                `jt`.`fichier_mpr` AS `mprFile`, `jt`.`generic_id` AS `genericId`, `jt`.`estampille` AS `timestamp`
+                `jt`.`fichier_mpr` AS `mprFile`, `jt`.`estampille` AS `timestamp`
             FROM `fabplan`.`job_type` AS `jt` WHERE `jt`.`id_job_type` = :id " . 
             (new \MYSQLDatabaseLockingReadTypes($databaseConnectionLockingReadType))->toLockingReadString() . ";"
         );
@@ -76,8 +78,11 @@ class JobType extends \ModelTypeGeneric implements \JsonSerializable
         
         if ($row = $stmt->fetch())
         {
+            $model = \Model::withID($db, $row["modelId"]);
+            $type = \Type::withImportNo($db, $row["typeNo"]);
+            
             $instance = new self(
-                $id, $row["jobId"], $row["modelId"], $row["typeNo"], $row["mprFile"], $row["genericId"], $row["timestamp"]
+                $id, $row["jobId"], $model, $type, $row["mprFile"], $row["timestamp"]
             );
         }
         else
@@ -91,11 +96,7 @@ class JobType extends \ModelTypeGeneric implements \JsonSerializable
             FROM `fabplan`.`job_type_params` AS `jtp` 
             INNER JOIN `fabplan`.`job_type` AS `jt` ON `jt`.`id_job_type` = `jtp`.`job_type_id`
             INNER JOIN `fabplan`.`door_types` AS `dt` ON `dt`.`importNo` = `jt`.`type_no`
-            INNER JOIN `fabplan`.`generics` AS `g` ON `g`.`id` = CASE 
-                WHEN `jt`.`generic_id` IS NOT NULL 
-                THEN `jt`.`generic_id` 
-                ELSE `dt`.`generic_id` 
-            END
+            INNER JOIN `fabplan`.`generics` AS `g` ON `g`.`id` = `dt`.`generic_id`
         	INNER JOIN `fabplan`.`generic_parameters` AS `gp` 
                 ON `gp`.`generic_id` = `g`.`id` AND `gp`.`parameter_key` = `jtp`.`param_key`
             WHERE `jt`.`id_job_type` = :id
@@ -107,7 +108,7 @@ class JobType extends \ModelTypeGeneric implements \JsonSerializable
         
         while($row = $stmt->fetch())	// Récupération des paramètres
         {
-            $parameter = (new JobTypeParameter($id, $row["param_key"], $row["param_value"]));
+            $parameter = (new \JobTypeParameter($id, $row["param_key"], $row["param_value"]));
             array_push($instance->_parameters, $parameter);
         }
         
@@ -127,34 +128,34 @@ class JobType extends \ModelTypeGeneric implements \JsonSerializable
             array_push($instance->_parts, $part);
         }
         
-        $this->setDatabaseConnectionLockingReadType($databaseConnectionLockingReadType);
+        $instance->setDatabaseConnectionLockingReadType($databaseConnectionLockingReadType);
         return $instance;
     }
     
     /**
      * Save the JobTypePorte object in the database
      *
-     * @param FabPlanConnection $db The database in which the record must be saved
+     * @param \FabPlanConnection $db The database in which the record must be saved
      *
      * @throws
      * @author Marc-Olivier Bazin-Maurice
      * @return JobTypePorte This JobtypePorte (for method chaining)
      */
-    public function save(FabPlanConnection $db) : JobType
+    public function save(\FabPlanConnection $db) : JobType
     {
-        if($this->getId() == null)
+        if($this->getId() === null || self::withID($db, $this->getId()) === null)
         {
             $this->insert($db);
         }
         else
         {
-            $dbTimestamp = \DateTime::createFromFormat("Y-m-d H:i:s", $this->getTimestampFromDatabase($db), "America/Montreal");
-            $localTimestamp = \DateTime::createFromFormat("Y-m-d H:i:s", $this->getTimestamp(), "America/Montreal");
-            if($this->getDatabaseConnectionReadingLockType() !== \MYSQLDatabaseLockingReadTypes::FOR_UPDATE)
+            $dbTimestamp = \DateTime::createFromFormat("Y-m-d H:i:s", $this->getTimestampFromDatabase($db));
+            $localTimestamp = \DateTime::createFromFormat("Y-m-d H:i:s", $this->getTimestamp());
+            if($this->getDatabaseConnectionLockingReadType() !== \MYSQLDatabaseLockingReadTypes::FOR_UPDATE)
             {
                 throw new \Exception("The provided " . get_class($this) . " is not locked for update.");
             }
-            elseif($databaseTimestamp > $localTimestamp)
+            elseif($dbTimestamp > $localTimestamp)
             {
                 throw new \Exception(
                     "The provided " . get_class($this) . " is outdated. The last modification date of the database entry is 
@@ -217,26 +218,26 @@ class JobType extends \ModelTypeGeneric implements \JsonSerializable
     private function insert(FabPlanConnection $db) : JobType
     {
         $stmt = $db->getConnection()->prepare("
-            INSERT INTO `fabplan`.`job_type` (`id_job_type`, `job_id`, `door_model_id`, `type_no`, `fichier_mpr`, `generic_id`)
-            VALUES (:jobTypeId, :jobId, :modelId, :typeNo, :mprFile, :genericId);
+            INSERT INTO `fabplan`.`job_type` (`id_job_type`, `job_id`, `door_model_id`, `type_no`, `fichier_mpr`)
+            VALUES (:jobTypeId, :jobId, :modelId, :typeNo, :mprFile);
         ");
         $stmt->bindValue(':jobTypeId', $this->getId(), PDO::PARAM_INT);
         $stmt->bindValue(':jobId', $this->getJobId(), PDO::PARAM_INT);
-        $stmt->bindValue(':modelId', $this->getModelId(), PDO::PARAM_INT);
-        $stmt->bindValue(':typeNo', $this->getTypeNo(), PDO::PARAM_INT);
+        $stmt->bindValue(':modelId', $this->getModel()->getId(), PDO::PARAM_INT);
+        $stmt->bindValue(':typeNo', $this->getType()->getImportNo(), PDO::PARAM_INT);
         $stmt->bindValue(':mprFile', $this->getMprFile(), PDO::PARAM_STR);
-        $stmt->bindValue(':genericId', $this->getGenericId(), PDO::PARAM_INT);
         $success = $stmt->execute();
         $this->setId($db->getConnection()->lastInsertId());
         
+        /* @var \JobTypeParameter $parameter */
         foreach($this->_parameters as $parameter)
         {
-            $parameter->save($db);
+            $parameter->setJobTypeId($this->getId())->save($db);
         }
         
         foreach($this->_parts as $part)
         {
-            $part->save($db);
+            $part->setJobTypeId($this->getId())->save($db);
         }
         
         return $this;
@@ -255,16 +256,14 @@ class JobType extends \ModelTypeGeneric implements \JsonSerializable
     {
         $stmt = $db->getConnection()->prepare("
             UPDATE `fabplan`.`job_type`
-            SET `job_id` = :jobId, `door_model_id` = :modelId, `type_no` = :typeNo, `fichier_mpr` = :mprFile, 
-                `generic_id`= :genericId
+            SET `job_id` = :jobId, `door_model_id` = :modelId, `type_no` = :typeNo, `fichier_mpr` = :mprFile
             WHERE `id_job_type` = :id;
         ");
         $stmt->bindValue(':id', $this->getId(), PDO::PARAM_INT);
         $stmt->bindValue(':jobId', $this->getJobId(), PDO::PARAM_INT);
-        $stmt->bindValue(':modelId', $this->getModelId(), PDO::PARAM_INT);
-        $stmt->bindValue(':typeNo', $this->getTypeNo(), PDO::PARAM_INT);
+        $stmt->bindValue(':modelId', $this->getModel()->getId(), PDO::PARAM_INT);
+        $stmt->bindValue(':typeNo', $this->getType()->getImportNo(), PDO::PARAM_INT);
         $stmt->bindValue(':mprFile', $this->getMprFile(), PDO::PARAM_STR);
-        $stmt->bindValue(':genericId', $this->getGenericId(), PDO::PARAM_INT);
         $success = $stmt->execute();
         
         foreach($this->_parameters as $parameter)
@@ -283,19 +282,26 @@ class JobType extends \ModelTypeGeneric implements \JsonSerializable
     /**
      * Delete the JobType object from the database
      *
-     * @param FabPlanConnection $db The database from which the record must be deleted
+     * @param \FabPlanConnection $db The database from which the record must be deleted
      *
      * @throws
      * @author Marc-Olivier Bazin-Maurice
-     * @return JobType This JobType (for method chaining)
+     * @return \JobType This JobType (for method chaining)
      */
-    public function delete(FabPlanConnection $db) : JobType
+    public function delete(\FabPlanConnection $db) : \JobType
     {
-        $this->emptyInDatabase($db);
-        
-        $stmt = $db->getConnection()->prepare("DELETE FROM `fabplan`.`job_type` WHERE `id_job_type` = :id;");
-        $stmt->bindValue(':id', $this->getId(), PDO::PARAM_INT);
-        $stmt->execute();
+        if($this->getDatabaseConnectionLockingReadType() !== \MYSQLDatabaseLockingReadTypes::FOR_UPDATE)
+        {
+            throw new \Exception("The provided " . get_class($this) . " is not locked for update.");
+        }
+        else
+        {
+            $this->emptyInDatabase($db);
+            
+            $stmt = $db->getConnection()->prepare("DELETE FROM `fabplan`.`job_type` WHERE `id_job_type` = :id;");
+            $stmt->bindValue(':id', $this->getId(), PDO::PARAM_INT);
+            $stmt->execute();
+        }
         
         return $this;
     }
